renamed views to pages
