+++ src/entry.js
@@ -5,13 +5,13 @@ import {
 } from 'react-router-dom';
 
 // Featured
-import About from './views/About';
-import Contact from './views/Contact';
-import Index from './views/Index';
-import NotFound from './views/NotFound';
-import Projects from './views/Projects';
-import Resume from './views/Resume';
-import Stats from './views/Stats';
+import About from './pages/About';
+import Contact from './pages/Contact';
+import Index from './pages/Index';
+import NotFound from './pages/NotFound';
+import Projects from './pages/Projects';
+import Resume from './pages/Resume';
+import Stats from './pages/Stats';
 
 // All of our CSS
 import './static/css/main.scss';
