+++ src/static/css/main.scss
@@ -60,9 +60,9 @@
 	@import 'layout/footer';
 	@import 'layout/menu';
 
-	@import 'views/about';
-	@import 'views/contact';
-	@import 'views/notFound';
-	@import 'views/resume';
-	@import 'views/skills';
-	@import 'views/stats';
+	@import 'pages/about';
+	@import 'pages/contact';
+	@import 'pages/notFound';
+	@import 'pages/resume';
+	@import 'pages/skills';
+	@import 'pages/stats';
