+++ src/static/css/pages/_notFound.scss
