+++ src/static/css/pages/_stats.scss
